+++ src/main/scala/com/lion/rafiki/sql/create.scala
@@ -6,7 +6,7 @@ import doobie.implicits._
 object create {
   val users =
     sql"""CREATE TABLE IF NOT EXISTS users (
-       id                 serial        PRIMARY KEY,
+       id                 bigserial        PRIMARY KEY,
        firstname          text,
        name               text,
        email              text          NOT NULL UNIQUE,
@@ -15,26 +15,26 @@ object create {
 
   val companies =
     sql"""CREATE TABLE IF NOT EXISTS companies (
-       id                 serial        PRIMARY KEY,
+       id                 bigserial        PRIMARY KEY,
        name               text          NOT NULL,
-       rh_user            integer       NOT NULL references users(id)
+       rh_user            bigint        NOT NULL references users(id)
     )""".update
 
   val companyContracts =
     sql"""CREATE TABLE IF NOT EXISTS company_contracts (
-       id                 serial        PRIMARY KEY,
-       company            integer       NOT NULL references companies(id),
-       kind               text          NOT NULL
+       id                 bigserial        PRIMARY KEY,
+       company            bigint           NOT NULL references companies(id),
+       kind               text             NOT NULL
     )""".update
 
   val forms =
     sql"""
     CREATE TABLE IF NOT EXISTS forms (
-       id                 serial        PRIMARY KEY,
-       company            integer       references companies(id),
-       name               text          NOT NULL,
+       id                 bigserial        PRIMARY KEY,
+       company            bigint           references companies(id),
+       name               text             NOT NULL,
        description        text,
-       tree_id            integer,
+       tree_id            bigint,
        tree_kind          form_tree_constr,
        FOREIGN KEY (tree_id, tree_kind)
           REFERENCES form_trees(id, kind)
@@ -52,9 +52,9 @@ object create {
     END $$$$;
 
     CREATE TABLE IF NOT EXISTS form_trees (
-       id                 serial        NOT NULL,
+       id                 bigserial        NOT NULL,
        kind               form_tree_constr NOT NULL,
-       parent_id          integer,
+       parent_id          bigint,
        parent_kind        form_tree_constr,
        PRIMARY KEY (id, kind),
        FOREIGN KEY (parent_id, parent_kind)
@@ -62,7 +62,7 @@ object create {
           ON DELETE CASCADE
     );
     CREATE TABLE IF NOT EXISTS form_tree_questions (
-       id INTEGER PRIMARY KEY,
+       id bigint PRIMARY KEY,
        kind
           form_tree_constr NOT NULL
           DEFAULT 'question'
@@ -76,7 +76,7 @@ object create {
           ON DELETE CASCADE
     );
     CREATE TABLE IF NOT EXISTS form_tree_groups (
-       id INTEGER PRIMARY KEY,
+       id bigint PRIMARY KEY,
        kind
           form_tree_constr NOT NULL
           DEFAULT 'group'
@@ -87,7 +87,7 @@ object create {
           ON DELETE CASCADE
     );
     CREATE TABLE IF NOT EXISTS form_tree_texts (
-       id INTEGER PRIMARY KEY,
+       id bigint PRIMARY KEY,
        kind
           form_tree_constr NOT NULL
           DEFAULT 'text'
